server/pyproject.toml
[tool.poetry]
name = "ethereum-indexer-graphql-server"
version = "0.1.0"
description = "Graphql server for blockchain indexer"
authors = ["RKL <hello@rl.games>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
tartiflette-aiohttp = "^1.4.1"
motor = "^2.5.1"
dacite = "^1.6.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pylint = "^2.13.3"
black = "^22.1.0"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
